+++ extensions/apidoc/CHANGELOG.md
@@ -1,8 +1,14 @@
 Yii Framework 2 apidoc extension Change Log
 ===========================================
 
-2.0.0-rc under development
---------------------------
+2.0.0 under development
+-----------------------
+
+- no changes in this release.
+
+
+2.0.0-rc September 27, 2014
+---------------------------
 
 - no changes in this release.